Actualizar entidades del modelo para estandarizar tipos de datos y añadir nuevas relaciones:
- Cambiar atributos `id` de tipo `int` a `Long` en varias entidades (`Producto`, `Categoria`, `MetodoPago`, etc.).
- Añadir nuevas entidades: `MovimientoInventario`.
- Renombrar clase `Imagen` a `ImagenProducto`.
- Configurar relaciones JPA en entidades (`StockSede`, `Categoria`, `Producto`).

+++ src/main/java/com/sistemaBackend/SistemaTittaBackend/model/Rol.java
@@ -12,7 +12,7 @@ public class Rol {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_rol")
-    private int idRol;
+    private Long idRol;
 
     @Column(name = "nombre_rol", nullable = false, unique = true, length = 50)
     private RolEnum nombreRol;
